requirements.txt
# Core
numpy>=1.24
opencv-python>=4.8
pandas>=2.0
tqdm>=4.66
PyYAML>=6.0

# Visualization / utilities
matplotlib>=3.7

# Detection (YOLO)
ultralytics>=8.1

# Jersey classifier training/inference
torchvision>=0.15
Pillow>=9.5

# Tracking
deep-sort-realtime>=1.3.2
filterpy>=1.4.5
scipy>=1.10

# Helpful for drawing/annotation pipelines (optional but practical)
supervision>=0.18
